requirements.txt
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
smtplib
imaplib
email
email.header
email.mime.text
email.mime.multipart
requests

